+++ sregistry/main/registry/push.py
@@ -54,11 +54,15 @@ def push(self, path, name, tag=None):
 
     # Extract the metadata
     names = parse_image_name(remove_uri(name), tag=tag)
-    metadata = self.get_metadata(path, names=names)
+    metadata = self.get_metadata(path, names=names) or {}
 
     # Add expected attributes
     if "data" not in metadata:
-        metadata['data'] = {'attributes': {'labels': dict() }}
+        metadata['data'] = {'attributes': {}}
+    if "labels" not in metadata['data']['attributes']:
+        metadata['data']['attributes']['labels'] = {}
+    if metadata['data']['attributes']['labels'] == None:
+        metadata['data']['attributes']['labels'] = {}
 
     # Try to add the size
     image_size = os.path.getsize(path) >> 20