+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The versions coincide with releases on pip. Only major versions will be released
 
 ## [0.0.x](https://github.com/singularityhub/sregistry-cli/tree/master) (0.0.x)
  - adding globus integration (0.0.80)
+ - fixing bug that registry does not have label for size in metadata (0.0.79)
  - consolidating shared pull functionality between nvidia and docker (0.0.78)
  - updating pull to handle whiteout files, adding THIRD-PARTY-NOTICES (0.0.77)
  - version bump to add google-compute builder startup files (0.0.76)